+++ pygama/dsp/old_dsp/calculators.py
