+++ pygama/dsp/old_dsp/dsp_base.py
