+++ pygama/dsp/processors/trap_filter.py
